+++ scripts/fix-imports.js
@@ -126,3 +126,7 @@ if (fs.existsSync(distDir)) {
 
 
 
+
+
+
+