+++ src/utils/calculations.ts
@@ -163,3 +163,7 @@ export function validateSkillValue(skillValue: number, masteryRank: number): num
 
 
 
+
+
+
+